ddl: handle store closed in doDDLJob (#18844)
* ddl: handle store closed in doDDLJob

* fix typo

* fix test

* fix test

* address comment

* address comments

* fix test

* address comment

* address comment

* fix data race

* fix data race

* don't repeat code

* address comment

* adress comment

* avoid failpoint change between two injection

* add comments

* change behaviour, fix test

* fix build option

* remove used variable

* Revert "remove used variable"

This reverts commit df93c431163515a496015f20f323ea91480f0020.

* modify comment

* try fix deadlock

* try fix deadlock

* try fix test
