+++ ddl/schema_test.go
@@ -26,7 +26,6 @@ import (
 	"github.com/pingcap/tidb/meta"
 	"github.com/pingcap/tidb/sessionctx"
 	"github.com/pingcap/tidb/types"
-	"github.com/pingcap/tidb/util/mock"
 )
 
 var _ = Suite(&testSchemaSuite{})
@@ -224,64 +223,6 @@ func (s *testSchemaSuite) TestSchemaWaitJob(c *C) {
 	c.Assert(err, IsNil)
 	schemaID := genIDs[0]
 	doDDLJobErr(c, schemaID, 0, model.ActionCreateSchema, []interface{}{dbInfo}, ctx, d2)
-}
-
-func testRunInterruptedJob(c *C, d *ddl, job *model.Job) {
-	ctx := mock.NewContext()
-	ctx.Store = d.store
-
-	done := make(chan error, 1)
-	go func() {
-		done <- d.doDDLJob(ctx, job)
-	}()
-
-	ticker := time.NewTicker(d.lease * 1)
-	defer ticker.Stop()
-LOOP:
-	for {
-		select {
-		case <-ticker.C:
-			d.Stop()
-			d.restartWorkers(context.Background())
-			time.Sleep(time.Millisecond * 20)
-		case err := <-done:
-			c.Assert(err, IsNil)
-			break LOOP
-		}
-	}
-}
-
-func (s *testSchemaSuite) TestSchemaResume(c *C) {
-	store := testCreateStore(c, "test_schema_resume")
-	defer store.Close()
-
-	d1 := testNewDDLAndStart(
-		context.Background(),
-		c,
-		WithStore(store),
-		WithLease(testLease),
-	)
-	defer d1.Stop()
-
-	testCheckOwner(c, d1, true)
-
-	dbInfo := testSchemaInfo(c, d1, "test")
-	job := &model.Job{
-		SchemaID:   dbInfo.ID,
-		Type:       model.ActionCreateSchema,
-		BinlogInfo: &model.HistoryInfo{},
-		Args:       []interface{}{dbInfo},
-	}
-	testRunInterruptedJob(c, d1, job)
-	testCheckSchemaState(c, d1, dbInfo, model.StatePublic)
-
-	job = &model.Job{
-		SchemaID:   dbInfo.ID,
-		Type:       model.ActionDropSchema,
-		BinlogInfo: &model.HistoryInfo{},
-	}
-	testRunInterruptedJob(c, d1, job)
-	testCheckSchemaState(c, d1, dbInfo, model.StateNone)
 }
 
 func testGetSchemaInfoWithError(d *ddl, schemaID int64) (*model.DBInfo, error) {
